refactor: separate out logcat and journalctl
