+++ uplink/src/base/mod.rs
@@ -3,8 +3,10 @@ use std::{collections::HashMap, fmt::Debug};
 
 use serde::{Deserialize, Serialize};
 
-#[cfg(any(target_os = "linux", target_os = "android"))]
-use crate::collector::logging::LoggerConfig;
+#[cfg(any(target_os = "linux"))]
+use crate::collector::journalctl::JournalCtlConfig;
+#[cfg(any(target_os = "android"))]
+use crate::collector::logcat::LogcatConfig;
 
 pub mod actions;
 pub mod bridge;
@@ -176,6 +178,8 @@ pub struct Config {
     pub action_redirections: HashMap<String, String>,
     #[serde(default)]
     pub ignore_actions_if_no_clients: bool,
-    #[cfg(any(target_os = "linux", target_os = "android"))]
-    pub logging: Option<LoggerConfig>,
+    #[cfg(any(target_os = "linux"))]
+    pub logging: Option<JournalCtlConfig>,
+    #[cfg(any(target_os = "android"))]
+    pub logging: Option<LogcatConfig>,
 }
